+++ cloudfunctions/post-detail/index.js
@@ -24,7 +24,7 @@ exports.main = wrap(async (event) => {
       const uid = me.data[0]._id
       const likeRes = await db.collection('likes').where({ userId: uid, targetId: postId, type: 'post' }).count()
       isLiked = likeRes.total > 0
-      const collectRes = await db.collection('collects').where({ userId: uid, targetId: postId }).count()
+      const collectRes = await db.collection('collects').where({ userId: uid, targetId: postId, type: 'post' }).count()
       isCollected = collectRes.total > 0
     }
   }